+++ tests/Serialization/MarcXmlTest.php
@@ -29,6 +29,8 @@
 
 namespace VuFindTest\Marc\Serialization;
 
+use VuFind\Marc\MarcReader;
+
 /**
  * MarcXml Test Class
  *
@@ -40,6 +42,8 @@ namespace VuFindTest\Marc\Serialization;
  */
 class MarcXmlTest extends \PHPUnit\Framework\TestCase
 {
+    use \VuFind\Marc\Test\Feature\FixtureTrait;
+
     /**
      * Test rewind
      *
@@ -62,5 +66,20 @@ class MarcXmlTest extends \PHPUnit\Framework\TestCase
         $this->expectExceptionMessage('Collection file not open');
         $class = new \VuFind\Marc\Serialization\MarcXml();
         $class->getNextRecord();
+    }
+
+    /**
+     * Test bad record
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testToStringException()
+    {
+        $xmlFile = $this->getFixture('marc/bad.mrc');
+        $reader = new MarcReader($xmlFile);
+        $this->expectException(\Exception::class);
+        $this->expectExceptionMessage('Error processing XML');
+        $xml = $reader->toFormat('MARCXML');
     }
 }